fixed forms issue #15
