+++ src/containers/App.tsx
@@ -1,8 +1,8 @@
 import './App.scss';
 import Header from '../components/Header/Header';
 import Footer from '../components/Footer/Footer';
-import Register from '../components/Register/Register';
-import Signin from '../components/Signin/Signin';
+import Register from '../components/Forms/Register/Register';
+import Signin from '../components/Forms/Signin/Signin';
 
 function App() {
   return (
